Fixed Picite and Exhibit issues. Fixes for SpaceBetweenSentences and InsertNBS as well.
LR Cites allow for pincites, Exhibits update correctly when pincites are added.

SpaceBetweenSentences: additional abbreviations and matching based on case. Known error: does not go through foot/endnotes because of time to complete.
NBS: added space in front of matched expressions. Known error: will not put nbs if expression is first word in a paragraph (needs to have a space in front of it).