Updated SpacesBetweenSentences to not do Range.Find and to apply to all StoryTypes
Update done because replacing some abbreviations caused loss of formatting. Updates just add a space to the end of each sentence if it does not ened with one of the abbreviations.

Need to add exceptions for colon, semicolon, parenthesis, slash, etc.
Need to fix so space is not added if there are already two spaces